+++ empire/server/api/app.py
@@ -157,7 +157,7 @@ def initialize(
     else:
         log.info("Starkiller disabled. Not loading.")
 
-    cert_path = os.path.abspath("./empire/server/data/")
+    cert_path = Path(empire_config.api.cert_path)
 
     if run:
         if not secure: